% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkandsummarize.R
\name{shrink_and_summarize}
\alias{shrink_and_summarize}
\title{Min-max Scaling on Numeric Values to a New Range and Compute Summary Statistics}
\usage{
shrink_and_summarize(x, lower_limit, upper_limit, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector to be scaled. Named 'x' because of the base R convention on function parameters.}

\item{lower_limit}{A numeric value specifying the minimum value for scaling, tells the user directly about the parameter}

\item{upper_limit}{A numeric value specifying the maximum value for scaling, tells the user directly about the parameter}

\item{na.rm}{Indicating whether NA values should be removed. Followed base R convention on NA values.}
}
\value{
A list containing two elements:
\item{prepared}{A numeric vector of the same length as input 'x', with values
scaled to the range. NA values in input remain NA in output.}
\item{summary}{A data frame with summary statistics (mean, median, sd) of the
scaled values.}
}
\description{
Scales a numeric vector to fit within a range using min-max scaling and compute summary (mean, median and std)
}
\examples{
x <- c(1, 2, 5, 50, 100, NA)
shrink_and_summarize(x, lower_limit = 0, upper_limit = 10)
y <- c(1, 2, 5, 50, 100, NA)
shrink_and_summarize(x, lower_limit = 0, upper_limit = 10, na.rm = TRUE)
}
